Moved FFText namespace to its own header file and cleaned up some unneeded inclusions.
